refactor: relocate Gitea loader dataclasses to types module
Motivation:
Placing logic-agnostic data structures in utils.py is architecturally
inconsistent. Moving them to a dedicated types module improves project
structure and maintainability.

Design Choices:
- Created openqabot/types/gitea.py for Gitea-specific type definitions.
- Relocated BuildTarget and RepoConfig dataclasses to the new module.
- Updated all call sites, tests, and utility functions to use the new
  location.
- Utilized TYPE_CHECKING blocks where appropriate to prevent circular
  imports and improve performance.

User Benefits:
- Better organized codebase.
- More predictable location for data structure definitions.
