+++ openqabot/utils.py
@@ -8,7 +8,6 @@ import logging
 import os
 import re
 from copy import deepcopy
-from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 from urllib.parse import urlparse
 
@@ -17,6 +16,7 @@ from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
 
 if TYPE_CHECKING:
+    from .types.gitea import BuildTarget, RepoConfig
     from .types.types import Data
 
 
@@ -128,25 +128,6 @@ def unique_dicts(dicts: list[dict[Any, Any]]) -> list[dict[Any, Any]]:
             seen.add(items)
             unique.append(d)
     return unique
-
-
-@dataclass
-class BuildTarget:
-    """Build target information."""
-
-    project: str
-    arch: str
-    product_name: str
-
-
-@dataclass
-class RepoConfig:
-    """Repository configuration parameters."""
-
-    repo_type: str
-    download_base_url: str
-    obs_download_url: str
-    obs_products: set[str] | None = None
 
 
 def get_repo_url(
